requirements.txt
streamlit
redis
pandas
openpyxl
python-dotenv
